Rename ::Test::HTTP::HashServer to ::Test::HashServer
Can't use "http" in Python code because otherwise interpreter tries to
import our own "http" instead of "http.client"
