+++ lib/MediaWords/Controller/Api/V2/t/Media.t
@@ -4,7 +4,7 @@ use warnings;
 use Modern::Perl '2015';
 use MediaWords::CommonLibs;
 
-use MediaWords::Test::HTTP::HashServer;
+use MediaWords::Test::HashServer;
 use Readonly;
 use Test::More;
 use Test::Deep;
@@ -117,12 +117,12 @@ sub test_for_scraped_feeds($$)
     }
 }
 
-# start a MediaWords::Test::HTTP::HashServer with the following pages for each of the given site names:
+# start a MediaWords::Test::HashServer with the following pages for each of the given site names:
 # * home page - http://localhost:$port/$site
 # * feed page - http://localhost:$port/$site/feed
 # * custom feed page - http://localhost:$port/$site/custom_feed
 #
-# return the MediaWords::Test::HTTP::HashServer instance
+# return the MediaWords::Test::HashServer instance
 sub _start_media_create_hash_server
 {
     my ( $site_names ) = @_;
@@ -154,7 +154,7 @@ ATOM
         $pages->{ "/$site_name/custom_feed" } = { header => 'Content-Type: application/atom+xml', content => $atom };
     }
 
-    my $hs = MediaWords::Test::HTTP::HashServer->new( $port, $pages );
+    my $hs = MediaWords::Test::HashServer->new( $port, $pages );
 
     $hs->start();
