+++ lib/MediaWords/Crawler/Download/Feed/t/feed_error.t
@@ -8,7 +8,7 @@ use MediaWords::CommonLibs;
 
 use Test::More tests => 2;
 
-use MediaWords::Test::HTTP::HashServer;
+use MediaWords::Test::HashServer;
 use Readonly;
 
 use MediaWords::Crawler::Engine;
@@ -60,7 +60,7 @@ sub test_invalid_feed($)
         '/foo' => '<rss version="2.0"><kim_kardashian></rss>',
     };
 
-    my $hs = MediaWords::Test::HTTP::HashServer->new( $HTTP_PORT, $pages );
+    my $hs = MediaWords::Test::HashServer->new( $HTTP_PORT, $pages );
 
     $hs->start;
