+++ lib/MediaWords/Test/TopicTweets.pm
@@ -14,7 +14,6 @@ use warnings;
 use Modern::Perl '2015';
 use MediaWords::CommonLibs;
 
-use MediaWords::Test::HTTP::HashServer;
 use Readonly;
 use Test::More;
 use File::Slurp;
