+++ lib/MediaWords/Util/URL/t/Variants.t
@@ -10,7 +10,7 @@ use Test::Deep;
 use Test::More tests => 15;
 
 use Readonly;
-use MediaWords::Test::HTTP::HashServer;
+use MediaWords::Test::HashServer;
 use HTTP::Status qw(:constants);
 use HTTP::Response;
 use Data::Dumper;
@@ -51,7 +51,7 @@ sub test_all_url_variants($)
         '/third'  => 'This is where the redirect chain should end.',
     };
 
-    my $hs = MediaWords::Test::HTTP::HashServer->new( $TEST_HTTP_SERVER_PORT, $pages );
+    my $hs = MediaWords::Test::HashServer->new( $TEST_HTTP_SERVER_PORT, $pages );
     $hs->start();
     @actual_url_variants = MediaWords::Util::URL::Variants::all_url_variants( $db, $starting_url );
     $hs->stop();
@@ -70,7 +70,7 @@ sub test_all_url_variants($)
         '/third'  => '<link rel="canonical" href="' . $TEST_HTTP_SERVER_URL . '/fourth" />',
     };
 
-    $hs = MediaWords::Test::HTTP::HashServer->new( $TEST_HTTP_SERVER_PORT, $pages );
+    $hs = MediaWords::Test::HashServer->new( $TEST_HTTP_SERVER_PORT, $pages );
     $hs->start();
     @actual_url_variants = MediaWords::Util::URL::Variants::all_url_variants( $db, $starting_url );
     $hs->stop();
@@ -93,7 +93,7 @@ sub test_all_url_variants($)
         '/second' => '<meta http-equiv="refresh" content="0; URL=/',
     };
 
-    $hs = MediaWords::Test::HTTP::HashServer->new( $TEST_HTTP_SERVER_PORT, $pages );
+    $hs = MediaWords::Test::HashServer->new( $TEST_HTTP_SERVER_PORT, $pages );
     $hs->start();
     @actual_url_variants = MediaWords::Util::URL::Variants::all_url_variants( $db, $starting_url );
     $hs->stop();
