+++ mediacloud/mediawords/annotator/test_cliff.py
@@ -2,7 +2,7 @@ import copy
 from typing import Union
 
 from mediawords.annotator.cliff import CLIFFAnnotator
-from mediawords.test.http.hash_server import HashServer
+from mediawords.test.hash_server import HashServer
 from mediawords.test.test_database import TestDatabaseWithSchemaTestCase
 from mediawords.util.config import get_config as py_get_config, set_config as py_set_config
 from mediawords.util.parse_json import encode_json
