+++ mediacloud/mediawords/util/url/test_variants.py
@@ -1,7 +1,7 @@
 import pytest
 
 from mediawords.test.db.create import create_test_story_stack, create_test_topic
-from mediawords.test.http.hash_server import HashServer
+from mediawords.test.hash_server import HashServer
 from mediawords.test.test_database import TestDatabaseWithSchemaTestCase
 from mediawords.util.network import random_unused_port
 from mediawords.util.url import urls_are_equal
