+++ t/test_crawler_auth.t
@@ -8,7 +8,7 @@ use MediaWords::CommonLibs;
 
 use English '-no_match_vars';
 
-use MediaWords::Test::HTTP::HashServer;
+use MediaWords::Test::HashServer;
 use Test::More tests => 6;
 
 use MediaWords::Crawler::Engine;
@@ -44,7 +44,7 @@ sub test_auth
     my $port = 8899;
     my $pages = { '/auth' => { auth => 'foo:bar', content => 'foo bar' } };
 
-    my $hs = MediaWords::Test::HTTP::HashServer->new( $port, $pages );
+    my $hs = MediaWords::Test::HashServer->new( $port, $pages );
 
     ok( $hs, 'hashserver object returned' );
