+++ t/test_tm_mine.t
@@ -14,7 +14,7 @@ use English '-no_match_vars';
 
 use Data::Dumper;
 use Digest::MD5 qw(md5_hex);
-use MediaWords::Test::HTTP::HashServer;
+use MediaWords::Test::HashServer;
 use Readonly;
 use Test::More;
 use Text::Lorem::More;
@@ -233,7 +233,7 @@ sub start_hash_servers
 
         map { $site_hash->{ $_->{ path } } = $_->{ content } } @{ $site->{ pages } };
 
-        my $hs = MediaWords::Test::HTTP::HashServer->new( $site->{ port }, $site_hash );
+        my $hs = MediaWords::Test::HashServer->new( $site->{ port }, $site_hash );
 
         DEBUG "starting hash server $site->{ id }";
